client/python: Improve message construction
Numpy's array2string function only returns a summary of the array when
it is larger than a threshold, resulting in something like this:

>>> x=np.arange(10000)
>>> np.array2string(x)
'[   0    1    2 ... 9997 9998 9999]'

This could be increased by passing the 'threshold' argument. Instead,
create an array of strings (one per value) and join them.

The use of f-strings requires Python 3.6.